fix(backtest): produce a valid equity curve and volatility-scaled stops
Final equity always equaled starting equity: the equity curve was
pre-filled flat and only updated on close bars, and every trade used a
fixed $28 stop that rarely triggered and never scaled across symbols.

- portfolio: seed equity curve as NaN + initial equity, then forward-fill
  so each realized-PnL step carries through to the last bar
- indicators: add ATR (rolling mean of True Range)
- engine: build a per-bar ATR series and size stops/targets from it
  instead of a hardcoded value; skip the entry bar when checking exits to
  remove same-bar look-ahead
- processor: sort formatted price data oldest to newest so ffill and
  indicators run in chronological order
- data_manager: return early when no date range is requested, avoiding a
  None comparison
- analysis: base win_percentage on closed trades with a divide-by-zero
  guard; export Analysis from the package

+++ src/backend/risk/sizing.py
@@ -2,7 +2,7 @@ class Sizing:
     '''Determines the position sizing'''
     
     @staticmethod
-    def volatility_targeted_sizing(risk_pct: float, equity: float, atr_multiplier: int, atr: int) -> int:
+    def volatility_targeted_sizing(risk_pct: float, equity: float, atr_multiplier: int, atr: float) -> int:
         '''Returns position size based on current market volatility'''
         dollar_risk = equity * risk_pct
         stop_distance = atr_multiplier * atr
